Http exercise 3 - pg. 63
Writing request body (post) to a file using steams and pipes (very easy).

To use streams just create destination steam (method of fs in case of files)
and call the sourcestream.pipe (and source is request which is also a stream).
The piping takes care of verything.

Experienced some issues described in code.